requirements.txt
requests
beautifulsoup4
lxml
pyotp
python-dotenv